Refactor component names and directory structure
