+++ src/components/Marketplace/MarketplaceNFTGrid/MarketplaceNFTCard/ListingModal/ListingModal.tsx
@@ -5,7 +5,7 @@ import { useContractFunction, useEthers } from "@usedapp/core";
 import { Modal, Box, Typography, CircularProgress } from "@mui/material";
 import { LoadingButton } from "@mui/lab";
 import ModalCloseButton from "./ModalCloseButton";
-import ModalPlayer from "./ModalPlayer";
+import ModalPlayer from "./ModalPlayer/ModalPlayer";
 
 const modalBoxStyle = {
   position: "absolute" as "absolute",
